moved setfavorite from event.gateway into user.gateway and tested using Postman
