+++ server/src/event/event.module.ts
@@ -1,5 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { ClientModule } from 'src/client/client.module';
+import { UserGateway } from 'src/user/user.gateway';
 import { AuthModule } from '../auth/auth.module';
 import { OrganizationModule } from '../organization/organization.module';
 import { PrismaModule } from '../prisma/prisma.module';
@@ -16,4 +17,4 @@ import { EventService } from './event.service';
   providers: [EventService, EventGateway],
   exports: [EventService, EventGateway],
 })
-export class EventModule {}
+export class EventModule { }
